# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FWA Sales Digitalization — Account Executive API
  version: 1.0.0-draft
  summary: Backend contract for the HiFi AIR Account Executive mobile application.
  description: |
    Backend API for the **Account Executive (AE)** module of the HiFi AIR Sales
    Management Platform.

    ## Scope
    AE activities only. GSE / DSE / DSF flows — including Open New Outlet (ONO)
    and Gadget Store sell-in — are explicitly **out of scope** for this contract.

    ## The AE flow this API serves
    1. **Login** with an HQ-issued AE code (case-insensitive).
    2. **Register a prospect** during door-to-door activity — name, phone,
       address, GPS, and an interaction status of `EDUKASI`, `HOT_LEADS`
       or `PURCHASE`.
    3. **Activate a customer** — scan the MSISDN barcode on the HiFi AIR box.
       The server resolves ICCID / IMEI / modem type from inventory; the client
       never supplies them.
    4. **Track performance** — activations, incentive, new customers, hot leads.

    ## Conventions
    * All timestamps are RFC 3339 with offset. The **server** is authoritative
      on time; `deviceTime` is recorded for audit but never trusted.
    * The AE identity always comes from the JWT. No endpoint accepts an
      `aeId` in a request body — a client cannot write on another AE's behalf.
    * Money is IDR in whole rupiah as an integer. The UI renders
      "Insentif 242" as hundreds of thousands; that scaling is presentation-only.
...

  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with AE code and password
      description: |
        AE credentials are created and distributed by IOH HQ. The AE code is
        matched **case-insensitively** — `AE-SIDOARJO1`, `ae-sidoarjo1` and
        `Ae-SidoArjo1` are the same account.
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [aeCode, password]
              properties:
                aeCode:
                  type: string
                  examples: ["AE-BENGKULU1"]
                password:
                  type: string
                  format: password
                deviceLabel:
                  type: string
                  description: Free-text device identifier, stored against the refresh token.
                  examples: ["Samsung A15 / android-14"]
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token
      security: []
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current refresh token
      operationId: logout
      responses:
        "204": { description: Logged out }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me:
    get:
      tags: [Auth]
      summary: Profile of the signed-in AE
      description: Supplies the read-only "ID AE" field shown on every input form.
      operationId: getMe
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountExecutive" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ===================================================================
  # CUSTOMERS
  # ===================================================================
...
  /activations:
    get:
      tags: [Activations]
      summary: List activations submitted by this AE
      description: |
        Backs **Daftar Aktivasi**. The `Activated` / `Not Activated` badge is
        the `status` field; `activationDate` is populated only once the Gross
        Add feed confirms the SIM went live.
      operationId: listActivations
      parameters:
        - $ref: "#/components/parameters/Period"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: "#/components/schemas/ActivationStatus" }
          style: form
          explode: true
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Paginated"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Activation" }
        "422": { $ref: "#/components/responses/DateRangeError" }

    post:
      tags: [Activations]
      summary: Submit a customer activation
      description: |
        Backs the **Aktivasi Pelanggan** form.

        The client sends only the customer, the scanned MSISDN, and the GPS fix.
        IMEI and modem type are resolved server-side from `fwa_inventory` and
        snapshotted onto the activation record — a client is never trusted to
        assert which device a MSISDN is bundled with.

        Created with `status: NOT_ACTIVATED`. The GA reconciliation job flips it
        to `ACTIVATED` and stamps `activationDate`.
      operationId: createActivation
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActivationCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activation" }
        "409":
          description: This MSISDN has already been activated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                alreadySold:
                  value:
                    error:
                      code: MSISDN_ALREADY_ACTIVATED
                      message: "MSISDN 6285882724305 was already activated on 2026-08-11."
        "422":
          description: |
            Validation failed — commonly `MSISDN_NOT_ALLOCATED` (the number is
            not in this AE's stock) or `CUSTOMER_NOT_OWNED`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /activations/{activationId}:
    get:
      tags: [Activations]
      summary: Retrieve one activation
      operationId: getActivation
      parameters:
        - name: activationId
          in: path
          required: true
          schema: { type: integer, format: int64, examples: [2987654321] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activation" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ===================================================================
  # ATTENDANCE
  # ===================================================================
  /attendance/check-in:
    post:
      tags: [Attendance]
      summary: Daily check-in with selfie and location
      description: |
        Multipart because of the mandatory selfie. Compress client-side to
        roughly 1024px / 70% quality before upload — raw phone photos fail on a
        field connection.

        A check-in outside the permitted geofence is **accepted** with
        `approvalStatus: PENDING_APPROVAL` and routed to CSE, matching the UI
        warning "Lokasi kamu berada di luar jangkauan".
      operationId: checkIn
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo, latitude, longitude]
              properties:
                photo:
                  type: string
                  format: binary
                latitude:  { type: number, format: double }
                longitude: { type: number, format: double }
                geoAccuracyM: { type: number, format: float }
                isMocked:  { type: boolean, default: false }
                note:      { type: string, description: "Catatan (Opsional)" }
                document:  { type: string, format: binary, description: "Dokumen (Opsional)" }
                deviceTime: { type: string, format: date-time }
      responses:
        "201":
          description: Checked in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendance" }
        "409":
          description: Already checked in today.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /attendance/check-out:
    post:
      tags: [Attendance]
      summary: Daily check-out
      operationId: checkOut
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude:  { type: number, format: double }
                longitude: { type: number, format: double }
                isMocked:  { type: boolean, default: false }
                deviceTime: { type: string, format: date-time }
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendance" }
        "409":
          description: No open check-in for today.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /attendance:
    get:
      tags: [Attendance]
      summary: Attendance history
      operationId: listAttendance
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Paginated"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Attendance" }
        "422": { $ref: "#/components/responses/DateRangeError" }

  # ===================================================================
  # REPORTS
  # ===================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token from `/auth/login`.
    serviceAuth:
      type: apiKey
      in: header
      name: X-Service-Key
      description: Shared secret for server-to-server calls.
  parameters:
    Period:
      name: period
      in: query
      description: Relative window. Ignored when `from`/`to` are supplied.
      schema:
        type: string
        enum: ["7d", "30d", "mtd", "ytd"]
        default: "7d"
    From:
      name: from
      in: query
      schema: { type: string, format: date }
    To:
      name: to
      in: query
      schema: { type: string, format: date }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: perPage
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Client-generated UUID. Replaying the same key with the same payload
        returns the original resource instead of creating a duplicate — this is
        what makes the mobile offline outbox safe to retry.
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing, expired, or invalid token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist, or does not belong to this AE.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    DateRangeError:
      description: |
        ADDED 2026-08-19. Every endpoint taking `from`/`to` returns
        `422 INVALID_DATE_RANGE` when `from` is later than `to`. The two
        parameters are individually valid dates, so nothing rejects the pair
        until the server compares them — which made this an undeclared status
        code on an otherwise schema-compliant request. Documented rather than
        silently swapped, because a reversed range is a client bug and quietly
        "fixing" it would hide one.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            reversedRange:
              value:
                error:
                  code: INVALID_DATE_RANGE
                  message: "The 'from' date must not be after the 'to' date."
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier. Never localised.
              examples: ["MSISDN_NOT_ALLOCATED"]
            message:
              type: string
              description: Human-readable, safe to show to the AE.
            details:
              type: array
              items:
                type: object
                properties:
                  field:  { type: string }
                  issue:  { type: string }
    Paginated:
      type: object
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          properties:
            page:       { type: integer }
            perPage:    { type: integer }
            total:      { type: integer }
            totalPages: { type: integer }
    AuthTokens:
      type: object
      properties:
        accessToken:  { type: string }
        refreshToken: { type: string }
        expiresIn:
          type: integer
          description: Access-token lifetime in seconds.
          examples: [900]
        mustChangePassword:
          type: boolean
          description: True on first login with an HQ-issued password.
        profile: { $ref: "#/components/schemas/AccountExecutive" }
    AccountExecutive:
      type: object
      properties:
        aeId:     { type: string, format: uuid }
        aeCode:   { type: string, examples: ["AE-BENGKULU1"] }
        fullName: { type: string, examples: ["Hendra Jaya"] }
        role:     { type: string, const: "AE" }
        region:
          type: object
          properties:
            regionCode: { type: string, examples: ["BENGKULU"] }
            regionName: { type: string, examples: ["Bengkulu"] }
        mpxCode:  { type: string }
        workShift:
          type: object
          properties:
            start: { type: string, examples: ["08:00"] }
            end:   { type: string, examples: ["18:00"] }
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, INACTIVE]
...
    GeoPoint:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:     { type: number, format: double, minimum: -90,  maximum: 90 }
        longitude:    { type: number, format: double, minimum: -180, maximum: 180 }
        accuracyM:    { type: number, format: float }
        verified:
          type: boolean
          description: Server-side check passed — accuracy within tolerance and not mocked.
        isMocked:
          type: boolean
          description: |
            Android reported a mock-location provider. Recorded and flagged for
            review, never used to silently reject a submission.
...
    ActivationStatus:
      type: string
      description: |
        * `NOT_ACTIVATED` — submitted, SIM not yet live (red badge)
        * `ACTIVATED` — Gross Add confirmed (green badge)
        * `FAILED` — GA feed reported activation failure
      enum: [NOT_ACTIVATED, ACTIVATED, FAILED, CANCELLED]
    ActivationCreate:
      type: object
      required: [customerId, msisdn, latitude, longitude]
      properties:
        customerId: { type: integer, format: int64, examples: [1234567890] }
        msisdn:
          type: string
          pattern: '^62[0-9]{8,13}$'
          description: Scanned from the box barcode, normalised to 62 form.
          examples: ["6285882724305"]
        latitude:   { type: number, format: double }
        longitude:  { type: number, format: double }
        geoAccuracyM: { type: number, format: float }
        isMocked:   { type: boolean, default: false }
        deviceTime: { type: string, format: date-time }
    Activation:
      type: object
      properties:
        activationId: { type: integer, format: int64, examples: [2987654321] }
        customer:
          type: object
          properties:
            customerId: { type: integer, format: int64, examples: [1234567890] }
            fullName:   { type: string, examples: ["Susanto"] }
        msisdn:      { type: string, examples: ["6289999999990"] }
        imei:        { type: string, examples: ["355806671396654"] }
        deviceModelCode: { type: string, examples: ["HKM 127+"] }
        geo:         { $ref: "#/components/schemas/GeoPoint" }
        submittedAt: { type: string, format: date-time }
        status:      { $ref: "#/components/schemas/ActivationStatus" }
        activationDate:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
          description: GA Date. Null until the Gross Add feed confirms.
    Attendance:
      type: object
      properties:
        attendanceId:   { type: string, format: uuid }
        attendanceDate: { type: string, format: date }
        checkInAt:      { type: string, format: date-time }
        checkInPhotoUrl: { type: string, format: uri }
        checkInGeo:     { $ref: "#/components/schemas/GeoPoint" }
        checkInAddress: { type: string }
        withinGeofence: { type: boolean }
        checkOutAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        note:           { type: string }
        documentUrl:    { type: string, format: uri }
        approvalStatus:
          type: string
          enum: [AUTO_APPROVED, PENDING_APPROVAL, APPROVED, REJECTED]
...